requirements.txt
psutil>=5.9.0
pystray>=0.19.0
Pillow>=9.0.0
pyinstaller>=6.0.0
